+++ src/components/Layout.tsx
@@ -4,7 +4,7 @@ import { SettingsIcon, QuestionOutlineIcon, InfoOutlineIcon } from "@chakra-ui/i
 import { PiDotsNineBold } from "react-icons/pi";
 import logo from "@/assets/google_meet_horizontal_wordmark_2020q4_2x_icon_124_40_292e71bcb52a56e2a9005164118f183b.png";
 import myLogo from "@/assets/IMG_0158_2_Large.jpeg";
-import SettingModal from "../sections/LandingPage/SettingModal";
+import SettingModal from "../sections/LandingPage/SettingModal.tsx";
 import { useState } from "react";
 
 export default function Layout() {
