+++ src/index.scss
@@ -1,11 +1,11 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700;800;900&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap");
 
 *,
 *::after,
 *::before {
   box-sizing: border-box;
-  font-family: "Menlo";
+  font-family: "Inter";
 }
